replace remaining camelcase url
